requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.2
cffi==1.17.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pefile==2023.2.7
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
setuptools==78.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
